using longs instead of ints for file size

+++ src/main/java/edu/odu/cs/cs350/ArchiveFile.java
@@ -5,8 +5,7 @@ import java.nio.file.Path;
 import edu.odu.cs.cs350.Enum.FileType;
 
 public class ArchiveFile extends WebsiteFile {
-	public ArchiveFile(int inSize, Path inPath) {
+	public ArchiveFile(long inSize, Path inPath) {
 		super(inSize, inPath, FileType.ARCHIVE);
 	}
-
 }
