+++ src/main/java/edu/odu/cs/cs350/NonCategoryFile.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import edu.odu.cs.cs350.Enum.FileType;
 
 public class NonCategoryFile extends WebsiteFile {	
-	public NonCategoryFile(int inSize, Path inPath) {
+	public NonCategoryFile(long inSize, Path inPath) {
 		super(inSize, inPath, FileType.NONCATEGORY);
 	}
